gather all fixtures into a central place
