+++ resotocore/resotocore/analytics/__init__.py
@@ -52,8 +52,6 @@ class AnalyticsEvent:
 
 
 class AnalyticsEventSender(ABC):
-    events: List[AnalyticsEvent]
-
     async def core_event(
         self, kind: str, context: Optional[Mapping[str, JsonElement]] = None, **counters: Union[int, float]
     ) -> AnalyticsEvent:
