+++ resotocore/tests/resotocore/analytics/__init__.py
@@ -1,8 +1,2 @@
-from pytest import fixture
-
-from resotocore.analytics import AnalyticsEventSender, NoEventSender, AnalyticsEvent, InMemoryEventSender
 
 
-@fixture
-def event_sender() -> InMemoryEventSender:
-    return InMemoryEventSender()
